Add xml config support
