+++ spring-social-google/spring-social-google/src/main/java/org/springframework/social/google/config/annotation/GoogleProviderConfigRegistrar.java
@@ -15,15 +15,10 @@
  */
 package org.springframework.social.google.config.annotation;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
-import org.springframework.social.UserIdSource;
 import org.springframework.social.config.annotation.AbstractProviderConfigRegistrarSupport;
-import org.springframework.social.config.xml.ApiHelper;
-import org.springframework.social.connect.Connection;
-import org.springframework.social.connect.UsersConnectionRepository;
 import org.springframework.social.google.api.Google;
+import org.springframework.social.google.config.GoogleApiHelper;
 import org.springframework.social.google.connect.GoogleConnectionFactory;
 import org.springframework.social.google.security.GoogleAuthenticationService;
 import org.springframework.social.security.provider.SocialAuthenticationService;
@@ -41,33 +36,6 @@ public class GoogleProviderConfigRegistrar extends AbstractProviderConfigRegistr
 	@Override
 	protected Class<? extends SocialAuthenticationService<?>> getAuthenticationServiceClass() {
 		return GoogleAuthenticationService.class;
-	}
-	
-	static class GoogleApiHelper implements ApiHelper<Google> {
-		
-		private final UsersConnectionRepository usersConnectionRepository;
-
-		private final UserIdSource userIdSource;
-
-		private GoogleApiHelper(UsersConnectionRepository usersConnectionRepository, UserIdSource userIdSource) {
-			this.usersConnectionRepository = usersConnectionRepository;
-			this.userIdSource = userIdSource;		
-		}
-
-		public Google getApi() {
-			if (logger.isDebugEnabled()) {
-				logger.debug("Getting API binding instance for Google provider");
-			}
-					
-			Connection<Google> connection = usersConnectionRepository.createConnectionRepository(userIdSource.getUserId()).findPrimaryConnection(Google.class);
-			if (logger.isDebugEnabled() && connection == null) {
-				logger.debug("No current connection; Returning default GoogleTemplate instance.");
-			}
-			return connection != null ? connection.getApi() : null;
-		}
-		
-		private final static Log logger = LogFactory.getLog(GoogleApiHelper.class);
-
 	}
 
 }
