+++ spring-social-google/spring-social-google/src/main/java/org/springframework/social/google/config/xml/GoogleNamespaceHandler.java
@@ -1,0 +1,13 @@
+package org.springframework.social.google.config.xml;
+
+import org.springframework.social.config.xml.AbstractProviderConfigBeanDefinitionParser;
+import org.springframework.social.config.xml.AbstractProviderConfigNamespaceHandler;
+
+public class GoogleNamespaceHandler extends AbstractProviderConfigNamespaceHandler {
+
+	@Override
+	protected AbstractProviderConfigBeanDefinitionParser getProviderConfigBeanDefinitionParser() {
+		return new GoogleConfigBeanDefinitionParser();
+	}
+
+}
