+++ spring-social-google/spring-social-google/src/main/resources/META-INF/spring.schemas
@@ -1,0 +1,2 @@
+http\://www.springframework.org/schema/social/spring-social-google.xsd=org/springframework/social/google/config/xml/spring-social-google-1.0.xsd
+http\://www.springframework.org/schema/social/spring-social-google-1.0.xsd=org/springframework/social/google/config/xml/spring-social-google-1.0.xsd
