+++ spring-social-google/spring-social-google/src/main/resources/org/springframework/social/google/config/xml/spring-social-google-1.0.xsd
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
+	xmlns:social="http://www.springframework.org/schema/social/google"
+	targetNamespace="http://www.springframework.org/schema/social/google">
+
+	<xs:element name="config">
+		<xs:complexType>
+			<xs:attribute name="app-id" use="required" type="xs:string" />
+			<xs:attribute name="app-secret" use="required" type="xs:string" />
+			<xs:attribute name="authentication-service-class" use="optional" type="xs:string" />
+		</xs:complexType>
+	</xs:element>
+
+</xs:schema>
